+ item, + marks

+++ src/main/java/com/example/demo/Repositories/ItemRepository.java
@@ -3,5 +3,5 @@ package com.example.demo.Repositories;
 import com.example.demo.Entity.Item;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ItemRepository extends CrudRepository<Item, Long> {
+public interface ItemRepository extends CrudRepository<Item, Integer> {
 }
